use np.int32 => np.int64

+++ cpp/include/utils/ioutils.hpp
@@ -48,7 +48,7 @@ class IoUtils {
   std::unique_ptr<CuSimLogger> logger_container_;
   std::unordered_map<std::string, int> word_idmap_, word_count_;
   std::vector<std::string> word_list_;
-  int num_lines_, remain_lines_;
+  int64_t num_lines_, remain_lines_;
 };  // class IoUtils
 
 } // namespace cusim
